deps/v8/test/inspector/debugger/wasm-step-a-lot-expected.txt

Tests repeated stepping through a large function (should not OOM)

Running test: test
Setting up global instance variable.
Got wasm script: wasm://wasm/e78ec4f6
Setting breakpoint
Paused 50 and running...
Paused 100 and running...
Paused 150 and running...
Paused 200 and running...
Paused 250 and running...
Paused 300 and running...
Paused 350 and running...
Paused 400 and running...
Paused 450 and running...
Paused 500 and running...
Paused 550 and running...
Paused 600 and running...
Paused 650 and running...
Paused 700 and running...
Paused 750 and running...
Paused 800 and running...
Paused 850 and running...
Paused 900 and running...
Paused 950 and running...
Paused 1000 and running...
test function returned.
Paused 1003 times.
